fix(stream)!: Shorten the crate name to anstream
Most `anstyle` crates are trying to emphasize that they integrate with
`anstyle` or are complementary to it.  `anstream` is a bit different in
that this is a top-level crate, independently usable outside of
`anstyle`, that the earliest beginners of Rust should use.

Shortening the name I think makes it easier to use and clarifies its
role.