fix(wincon): Always reset at end of write
A lot of the complexity in designing `anstyle-wincon` comes back to the
state tracking inside of it.  This state tracking is to help reduce
redundant calls.  The primary user of this package is only calling it
when the state changes.  However, this does still have a side effect of
back-to-back styles having an extra color write + flush.  In at least my
cases, this isn't too often and ... it is just legacy windows.