feat(stream): Query the current choice
Have a situation in cargo where we use `AutoStream::new` and then need
to later check the color choice.  Rather than requiring the caller to
check both values and `and` them together, let's just report it.

Ideally, we could remove those color choice checks but that will take a
bit of plumbing work.