fix(svg): Allow setting min width
Before, we hard coded it if there was no content but allowed anything if
there was.  The problem without one is that it can be harder to browse a
list of diffs in github as they vary in size.

Now, we consistently set a minimum.
This shouldn't interefere with people getting the "real" output of the
command as the `text` area remains unchanged.