chore(ci): Run more checks on Windows
Because we have `anstyle-wincon`

MSRV is important for holding up our guarentees.

I skipped minimal versions because of some problem with `winapi` that I
couldn't decipher and the community is less interested in those being
correct.

I didn't add clippy as I was unsure how it would work out for checks
that are duplicate across platforms.