Extend `anstyle-svg` API to make it easier to integrate it into bigger HTML pages (#261)
I'm using this crate within `triagebot` for syntax highlighting of the
logs. We have extra content we need to add in the HTML, meaning I can't
simply use `render_html` or I'd need to parse the HTML, which would be
suboptimal. So instead, I keep the two parts that I need: the style and
the HTML "body".