feat(anstream): Provide read-only access to inner stream
The main concern we have is `&mut Inner` because our processing of input
is stateful and this can mess things up.
Adding an `as_inner` doesn't have that problem and can be useful for any
custom behavior users have added that they want to query.
This particularly came up with converting Rustup from termcolor to
anstream because of their test stream.